Code/Client/client/requirements.txt
astroid==2.4.1
bash==0.6
certifi==2019.11.28
cffi==1.14.0
chardet==3.0.4
colorama==0.4.3
cryptography==2.8
ecdsa==0.15
idna==2.9
iso8601==0.1.12
isort==4.3.21
lazy-object-proxy==1.4.3
mccabe==0.6.1
pip-install-test==0.5
pyasn1==0.4.8
pycparser==2.20
pylint==2.5.2
PyNaCl==1.3.0
python-dateutil==2.8.1
python-jose==3.1.0
requests==2.23.0
rsa==4.0
securesystemslib==0.14.2
six==1.14.0
toml==0.10.1
tuf==0.12.2
typed-ast==1.4.1
urllib3==1.25.8
wrapt==1.12.1
zipfile36==0.1.3
